requirements.txt
torch
torchvision
fastapi
uvicorn
pillow
minio
python-multipart
